source/spark_processor/requirements.txt
pyspark==3.5.0
delta-spark==3.1.0
confluent-kafka==2.3.0
boto3==1.34.14
python-dotenv==1.0.0
